+++ lib/http2-insecure.ts
@@ -38,7 +38,8 @@ export class Http2Server extends net.Server implements base.Http2Server {
         p1?: string | number,
         p2?: number
     ): Promise<AddressInfo> {
-        return new Promise(resolve => {
+        return new Promise((resolve, reject) => {
+            this.on('error', err => reject(err))
             this.listen(port, p1 as any, p2, () =>
                 resolve(this.address() as AddressInfo)
             )