+++ lib/http2-secure.ts
@@ -44,7 +44,8 @@ export class Http2SecureServer extends tls.Server
         p1?: string | number,
         p2?: number
     ): Promise<AddressInfo> {
-        return new Promise(resolve => {
+        return new Promise((resolve, reject) => {
+            this.on('error', err => reject(err))
             this.listen(port, p1 as any, p2, () =>
                 resolve(this.address() as AddressInfo)
             )