+++ lib/https.ts
@@ -43,7 +43,8 @@ export class HttpsServer extends base.Server {
         p1?: string | number,
         p2?: number
     ): Promise<AddressInfo> {
-        return new Promise(resolve => {
+        return new Promise((resolve, reject) => {
+            this.on('error', err => reject(err))
             this.listen(port, p1 as any, p2, () =>
                 resolve(this.address() as AddressInfo)
             )