+++ test/test-https.ts
@@ -18,7 +18,7 @@ test('close()', async t => {
     let server = createHttpsServer({}, () => {})
 
     await server.listenAsync()
-    await new Promise(resolve => {
+    await new Promise<void>(resolve => {
         server.close((err?: Error) => {
             t.ok(!err, 'shall call close-callback without an error')
             resolve()
@@ -38,10 +38,25 @@ test('close without listening', async t => {
 
     server = createHttpsServer({}, () => {})
 
-    await new Promise(resolve => {
+    await new Promise<void>(resolve => {
         server.close((err?: Error) => {
             t.ok(err, 'shall call close-callback with an error')
             resolve()
         })
     })
+})
+
+test('throw an error if listen fails', async t => {
+    const server1 = createHttpsServer({}, () => {})
+    const server2 = createHttpsServer({}, () => {})
+
+    await server1.listenAsync(12345)
+    try {
+        await server2.listenAsync(12345)
+        t.fail('shall throw an error if listen fails')
+    } catch (err) {
+        t.pass('shall throw an error if listen fails')
+    }
+
+    await server1.closeAsync()
 })