+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Support listening on unix sockets
+- Throw an error if listen fails
 
 ## [1.1.2] - 2020-01-10